refactor(materializer): extract PatchedReadonlyMap (file #3)
Lift the generic copy-on-read ReadonlyMap view + createPatchedReadonlyMap
factory into patched-readonly-map.ts (shared index-fast-path infra). Net 74/74, check 0.
